update fct report generation
