+++ legend-engine-xts-analytics/legend-engine-xts-analytics-lineage/legend-engine-xt-analytics-lineage-pure/src/test/java/org/finos/legend/pure/code/core/Test_Analytics_Lineage_FCT.java
@@ -17,20 +17,14 @@ package org.finos.legend.pure.code.core;
 import junit.framework.Test;
 import junit.framework.TestSuite;
 import org.eclipse.collections.api.factory.Lists;
-import org.eclipse.collections.api.list.ImmutableList;
-import org.eclipse.collections.api.list.MutableList;
 import org.eclipse.collections.api.map.MutableMap;
 import org.eclipse.collections.impl.factory.Maps;
-import org.finos.legend.engine.test.fct.FCTReport;
-import org.finos.legend.engine.test.fct.FCTTestCollection;
 import org.finos.legend.engine.test.fct.FCTTestSuitBuilder;
-import org.finos.legend.pure.code.core.relational.Test_Pure_Relational_FCT_Collection;
-import org.finos.legend.pure.m3.pct.reports.config.exclusion.ExclusionSpecification;
 import org.finos.legend.pure.runtime.java.compiled.execution.CompiledExecutionSupport;
 import org.finos.legend.pure.runtime.java.compiled.testHelper.PureTestBuilderCompiled;
 import static org.finos.legend.engine.test.shared.framework.PureTestHelperFramework.*;
 
-public class Test_Analytics_Lineage_FCT extends FCTReport
+public class Test_Analytics_Lineage_FCT  extends LineageFCTReport
 {
 
     public static Test suite()
@@ -46,35 +40,5 @@ public class Test_Analytics_Lineage_FCT extends FCTReport
                 () -> false,
                 Lists.mutable.empty()
         );
-    }
-
-    private static ImmutableList<FCTTestCollection> testCollection(CompiledExecutionSupport support)
-    {      return Lists.immutable.with(
-                Test_Pure_Relational_FCT_Collection.buildCollection(support));
-    }
-
-
-    @Override
-    public  ImmutableList<FCTTestCollection> getTestCollection()
-    {
-        return testCollection(PureTestBuilderCompiled.getClassLoaderExecutionSupport());
-    }
-
-    @Override
-    public String getreportID()
-    {
-        return "LineageFCT";
-    }
-
-    @Override
-    public String getStoreID()
-    {
-        return "relationalStore";
-    }
-
-    @Override
-    public MutableList<ExclusionSpecification> expectedFailures()
-    {
-        return Lists.mutable.empty();
     }
 }
